feat(navbar): implement senior dev corrections
Applied corrections from the senior developer on Wednesday, 20th August:

- Replaced button usage with `Link` by converting the reusable button component into a reusable aria-link component.
- Updated SVGs from Figma by using React Icons.
- Fixed typo in `scrollBehavior`.
