+++ src/app/globals.css
@@ -39,5 +39,4 @@ body {
   padding: 0;
   margin: 0;
   box-sizing: border-box;
-  behavior: scroll;
 }
