+++ src/components/Home/Header.js
@@ -1,7 +1,6 @@
-import Link from "next/link";
 
 import Navbar from "@/components/Navbar";
-import Button from "@/components/UI/Button";
+import Link from "@/components/UI/Link";
 import Logo from "@/components/UI/Logo";
 import HamburgerMenu from "@/components/MenuTrigger";
 
@@ -15,8 +14,8 @@ const Header = () => {
         <HamburgerMenu />
 
         <div className="@4xl:flex hidden gap-4 font-medium text-[1rem] ">
-          <Button className='py-[0.5rem] px-[2.5rem]'><Link href='/auth/signup'>Sign up</Link></Button>
-          <Button className='py-[0.5rem] px-[2.5rem]' variant="secondary"><Link href='/auth/signin'> Login</Link></Button>
+          <Link className='py-2.5 px-10' href='/auth/signup'>Sign up</Link>
+          <Link className='py-2.5 px-10' variant='secondary' href='/auth/signin'> Login</Link>
         </div>
       </div>
